+++ mobile-app/lib/pages/top_level/kennel_list_controller.dart
@@ -481,6 +481,7 @@ class KennelsListPageController extends GetxController {
       KennelFeature.enterKennelAdmin,
       appAccessFlags: kennel.hkm?.appAccessFlags ?? 0,
       mismanagementRoles: kennel.hkm?.mismanagementRoles ?? 0,
+      kennelOverrideJson: kennel.kennel.permissionOverrideJson,
     )) {
       showHcSnackbar(
         'You don\'t have admin access for this kennel.',